Add DXGI-exception handling

+++ V-Engine/App.cpp
@@ -2,7 +2,7 @@
 
 App::App()
 	:
-	m_wnd(800, 600, "V-Engine")
+	m_wnd( 800, 600, "V-Engine" )
 {
 }
 
@@ -21,7 +21,7 @@ int App::Run()
 
 void App::DoFrame()
 {
-	const float c = sin(m_clock.PeekDeltatime().asSeconds()) / 2.0f + 0.5;
-	m_wnd.gfx.ClearBuffer(c, c, 1.0f);
+	const double c = sin( m_clock.PeekDeltatime().asSeconds() ) / 2.0 + 0.5;
+	m_wnd.gfx.ClearBuffer( c, c, 1.0f );
 	m_wnd.gfx.EndFrame();
 }
